feat: modify ethereum configuration related functions
